# contento.web/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi documentación de CONTENTO",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "a57dc73b1f4c27874649b9650d81fe5d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@mail.com",
                                    "password": "12345"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario logueado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/readCsf": {
            "post": {
                "tags": [
                    "Constancia"
                ],
                "operationId": "e286d1703e67e2d91ccab53a641ecc7d",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "constancia": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "constancia": "file"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sendticket": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "operationId": "4e5634eacb38340048dc6a33ae022594",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codigo": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "codigo": "1234567890"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ticket": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "operationId": "19cd991078acd1cae5ae842157a6b689",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "forma_pago": {
                                        "type": "string"
                                    },
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "contacto": {
                                        "type": "string"
                                    },
                                    "referencia": {
                                        "type": "string"
                                    },
                                    "cantidad": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "forma_pago": "Efectivo",
                                    "tipo": "Whatsapp",
                                    "contacto": "3141335870",
                                    "referencia": "referencia",
                                    "cantidad": "50.00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/modifyticket/{codigo}": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "operationId": "effa49de6a02f1aa745469f92fe2ff42",
                "parameters": [
                    {
                        "parameter": "codigo",
                        "name": "codigo",
                        "in": "query",
                        "description": "El código del Ticket",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "forma_pago": {
                                        "type": "string"
                                    },
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "contacto": {
                                        "type": "string"
                                    },
                                    "referencia": {
                                        "type": "string"
                                    },
                                    "cantidad": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "forma_pago": "Efectivo",
                                    "tipo": "Whatsapp",
                                    "contacto": "3141335870",
                                    "referencia": "referencia",
                                    "cantidad": "50.00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    }
}